requirements.txt
fastapi==0.115.9
uvicorn==0.34.3
streamlit==1.43.2
langchain==0.3.25
langchain-chroma==0.2.4
langchain-huggingface==0.2.0
langchain-deepseek==0.1.3
chromadb==1.0.12
sentence-transformers==4.1.0
huggingface-hub==0.32.4
pydantic==2.11.7
python-dotenv==1.0.1
requests==2.32.3
